pour windows

+++ Makefile
@@ -1,25 +1,32 @@
-EXEC = build/Akropolis 
+OS := $(shell uname 2>/dev/null || echo Windows)
 
-# fichiers sources
+EXEC = build/Akropolis
+
 SRC = src/Hexagone.cpp src/main.cpp src/Participant.cpp src/Partie.cpp src/Plateau.cpp src/Tuile.cpp
-
 OBJ = $(patsubst src/%.cpp, build/%.o, $(SRC))
 
-# Options
 CXX = g++
-CXXFLAGS = -std=c++20 -Wall -Wextra -g -Iinclude -finput-charset=UTF-8 -fexec-charset=UTF-8
+CXXFLAGS = -std=c++20 -Wall -Wextra -g -Iinclude
 
 all: build $(EXEC)
 
 build/%.o: src/%.cpp | build
 	$(CXX) $(CXXFLAGS) -c $< -o $@
 
-# Création executable
 $(EXEC): $(OBJ)
-	$(CXX) $(CXXFLAGS) $(OBJ) -o $(EXEC)
+	$(CXX) $(CXXFLAGS) $(OBJ) -o $(EXEC)$(EXT)
 
 build:
+ifeq ($(OS),Windows)
+	if not exist build mkdir build
+else
 	mkdir -p build
+endif
 
 clean:
+ifeq ($(OS),Windows)
+	if exist build rmdir /S /Q build
+else
 	rm -rf build
+endif
+
